csn1: Fix an infinite loop in CSN.1 dissector when having more than 255 padding bits
Port of wireshark.git 8b5aa913711b32b1e1bc707919d2a98c1875d443.

Ported-by: Pau Espin Pedrol <pespin@sysmocom.de>

+++ src/csn1.cpp
@@ -1150,7 +1150,7 @@ csnStreamDecoder(csnStream_t* ar, const CSN_DESCR* pDescr, bitvec *vector, unsig
         {
           while (remaining_bits_len > 0)
           {
-            guint8 bits_to_handle = remaining_bits_len%8;
+            guint bits_to_handle = remaining_bits_len%8;
             if (bits_to_handle > 0)
             {
 	      LOGPC(DCSN1, LOGL_NOTICE, "%" PRIu64 "|", bitvec_read_field(vector, &readIndex, bits_to_handle));
